backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
cloudpickle==3.1.2
demucs==4.0.1
distro==1.9.0
docstring_parser==0.17.0
dora_search==0.1.12
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.128.8
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
julius==0.2.7
lameenc==1.8.1
llvmlite==0.43.0
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
omegaconf==2.3.0
openai-whisper==20250625
openunmix==1.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-multipart==0.0.20
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
retrying==1.4.2
sniffio==1.3.1
starlette==0.49.3
submitit==1.5.4
sympy==1.14.0
tiktoken==0.12.0
torch==2.8.0
torchaudio==2.8.0
tqdm==4.67.3
treetable==0.2.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
